camera sensor bugfixing (reference to osghandle) twowheeled robot added (nimm2 with camera) sense function added to robots (before control): sensors (type Sensor) are checked here position and optical flow camera sensors added

+++ ode_robots/robots/arm.cpp
@@ -462,7 +462,8 @@ namespace lpzrobots{
     joints.push_back(FJ_hand);		
  
 		
-    printf("size: %ld objects, %ld joints, %ld hingeservos\n", objects.size(), joints.size(), hingeServos.size());
+    printf("size: %ld objects, %ld joints, %ld hingeservos\n", (long int)objects.size(), 
+	   (long int)joints.size(), (long int)hingeServos.size());
     
     FOREACH(list<Sensor*>, conf.sensors, i){
 	(*i)->init(objects[hand]);
